NEW: Upgrade to SilverStripe 4 * Update composer.json * More updates to templates, missing libraries, etc * update readme + docs, require silverstripe/framework ^4 * convert to vendormodule, replace abandoned package * GridFieldExtensions is actually optional, move to suggests * use $table_name to ensure database tables remain unchanged * Rename folder structure for PSR-4 compliance
