+++ README.md
@@ -1,1 +1,14 @@
 # Silverstripe helpdesk #
+
+1. Update your composer.json:
+```
+    "require": {
+        "codecoyotes/silverstripe-helpdesk": "^1"
+    },
+    "repositories": {
+        "codecoyotes/silverstripe-helpdesk": {
+            "type": "vcs",
+            "url": "git@github.com:codecoyotes/silverstripe-helpdesk.git"
+        }
+    },
+```
