+++ composer.json
@@ -1,7 +1,7 @@
 {
   "name": "codecoyotes/silverstripe-helpdesk",
-  "description": "Helpdesk module for silverstripe",
-  "type": "silverstripe-module",
+  "description": "Helpdesk module for Silverstripe",
+  "type": "silverstripe-vendormodule",
   "keywords": [
 	"silverstripe",
 	"helpdesk",
@@ -17,7 +17,11 @@
 	"installer-name": "helpdesk"
   },
   "require": {
-	"silverstripe/tagfield": "1.3.2",
-	"silverstripe-australia/gridfieldextensions": "1.4.2"
+    "silverstripe/framework": "^4",
+	"silverstripe/tagfield": "^2"
+  },
+  "suggest": {
+      "symbiote/silverstripe-gridfieldextensions": "Allows drag-and-drop ordering of FAQ items in the CMS",
+      "silverstripe/spamprotection": "Spam protection for your helpdesk tickets"
   }
 }
