+++ src/Controllers/FaqPageController.php
@@ -1,0 +1,5 @@
+<?php
+namespace CodeCoyotes\Helpdesk;
+class FaqPageController extends \PageController {
+
+}
