+++ templates/CodeCoyotes/Helpdesk/Email/NewTicketCommentEmail.ss
