+++ templates/CodeCoyotes/Helpdesk/Email/NewTicketEmail.ss
